+++ src/indexes.ts
@@ -1,27 +1,28 @@
-import IndexResource from "./workflow/indexResource";
+export const JS_INDEX_URL =
+    "https://github.com/vemel/github_actions_js/tree/{ref}/nodejs_workflows";
+export const PY_INDEX_URL =
+    "https://github.com/vemel/github_actions_js/tree/{ref}/python_workflows";
 
-export const REPO_URL =
-    "https://raw.githubusercontent.com/vemel/github_actions_js";
-export const JS_INDEX_URL = `${REPO_URL}/{ref}/workflows/index.yml`;
-export const PY_INDEX_URL = `${REPO_URL}/{ref}/workflows_py/index.yml`;
+interface IIndex {
+    url: string;
+    shortcut?: string;
+}
 
-export const INDEXES: Array<IndexResource> = [
-    new IndexResource(
-        "GitHub Actions for Node.js projects",
-        JS_INDEX_URL,
-        "node",
-        "package.json"
-    ),
-    new IndexResource(
-        "GitHub Actions for Python projects",
-        PY_INDEX_URL,
-        "python",
-        "setup.py"
-    )
+export const INDEXES: Array<IIndex> = [
+    {
+        url: JS_INDEX_URL,
+        shortcut: "node"
+    },
+    {
+        url: PY_INDEX_URL,
+        shortcut: "python"
+    }
 ];
 
-export function getIndexResource(id: string): IndexResource {
-    const result = INDEXES.find(index => index.id === id || index.url === id);
-    if (!result) return new IndexResource(id, id, id);
-    return result;
+export function getIndexResource(url: string): IIndex {
+    const result = INDEXES.find(
+        index => index.url === url || index.shortcut === url
+    );
+    if (result) return result;
+    return { url };
 }